+++ packages/waallet/background/provider.ts
@@ -3,9 +3,9 @@ import * as ethers from "ethers"
 import { type Account } from "~packages/account"
 import { UserOperation } from "~packages/bundler"
 import { BundlerProvider } from "~packages/bundler/provider"
-import { JsonRpcProvider } from "~packages/jsonrpc/provider"
 import { type Paymaster } from "~packages/paymaster"
 import { NullPaymaster } from "~packages/paymaster/NullPaymaster"
+import { JsonRpcProvider } from "~packages/rpc/json/provider"
 import number from "~packages/util/number"
 import type { BigNumberish, HexString } from "~typing"
